Final upload. Three microservices completed to perform tasks identified in the README.md. Began completion of the Orchestrator.

+++ autoloan/src/main/java/com/example/banking/controller/AutoLoanController.java
@@ -1,6 +1,7 @@
 package com.example.banking.controller;
 
 import com.example.banking.models.AutoLoan;
+import com.example.banking.models.DefaultMessage;
 import com.example.banking.service.AutoLoanService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -27,24 +28,28 @@ public class AutoLoanController {
     // Read
     @GetMapping(value = "/getLoansByClientId/{clientId}", produces = "application/json")
     public ResponseEntity<List<AutoLoan>> getLoansByClientId(@PathVariable("clientId") int clientId) {
+        if (this.service.databaseCheck()) { return new ResponseEntity(new DefaultMessage(), HttpStatus.NOT_FOUND); }
         return new ResponseEntity(this.service.getLoansByClientId(clientId), HttpStatus.valueOf(200));
     }
 
     // Read All
     @GetMapping(value = "/getAllLoans", produces = "application/json")
     public ResponseEntity<List<AutoLoan>> getAllLoans() {
+        if (this.service.databaseCheck()) { return new ResponseEntity(new DefaultMessage(), HttpStatus.NOT_FOUND); }
         return new ResponseEntity(this.service.getAllLoans(), HttpStatus.valueOf(200));
     }
 
     // Update
     @PutMapping(value = "/updateLoan/{id}", produces = "application/json")
     public ResponseEntity<AutoLoan> updateLoan(@PathVariable("id") int id, @RequestBody AutoLoan updateAutoLoan) {
+        if (this.service.databaseCheck()) { return new ResponseEntity(new DefaultMessage(), HttpStatus.NOT_FOUND); }
         return new ResponseEntity(this.service.updateLoan(id, updateAutoLoan), HttpStatus.valueOf(200));
     }
 
     // Delete
     @DeleteMapping(value = "/deleteLoan/{id}", produces = "application/json")
     public ResponseEntity<Boolean> deleteLoan(@PathVariable("id") int id) {
+        if (this.service.databaseCheck()) { return new ResponseEntity(new DefaultMessage(), HttpStatus.NOT_FOUND); }
         if(this.service.delete(id)) {
             return new ResponseEntity(HttpStatus.valueOf(200));
         } else {
