+++ deposit/src/main/java/com/example/banking/controller/DepositController.java
@@ -1,5 +1,6 @@
 package com.example.banking.controller;
 
+import com.example.banking.models.DefaultMessage;
 import com.example.banking.models.Deposit;
 import com.example.banking.models.NewDepositAccount;
 import com.example.banking.service.DepositService;
@@ -28,24 +29,28 @@ public class DepositController {
     // Read
     @GetMapping(value = "/getDepositAccountsByClientId/{clientId}", produces = "application/json")
     public ResponseEntity<List<Deposit>> getDepositAccountsByClientId(@PathVariable("clientId") int clientId) {
+        if (this.service.databaseCheck()) { return new ResponseEntity(new DefaultMessage(), HttpStatus.NOT_FOUND); }
         return new ResponseEntity(this.service.getDepositAccountsByClientId(clientId), HttpStatus.valueOf(200));
     }
 
     // Read All
     @GetMapping(value = "/getAllDepositAccounts", produces = "application/json")
     public ResponseEntity<List<Deposit>> getAllDepositAccounts() {
+        if (this.service.databaseCheck()) { return new ResponseEntity(new DefaultMessage(), HttpStatus.NOT_FOUND); }
         return new ResponseEntity(this.service.getAllDepositAccounts(), HttpStatus.valueOf(200));
     }
 
     // Update
     @PutMapping(value = "/updateDepositAccount/{id}", produces = "application/json")
     public ResponseEntity<Deposit> updateLoan(@PathVariable("id") int id, @RequestBody Deposit updateDeposit) {
+        if (this.service.databaseCheck()) { return new ResponseEntity(new DefaultMessage(), HttpStatus.NOT_FOUND); }
         return new ResponseEntity(this.service.updateDeposit(id, updateDeposit), HttpStatus.valueOf(200));
     }
 
     // Delete
     @DeleteMapping(value = "/deleteDepositAccount/{id}", produces = "application/json")
     public ResponseEntity<Boolean> deleteDepositAccount(@PathVariable("id") int id) {
+        if (this.service.databaseCheck()) { return new ResponseEntity(new DefaultMessage(), HttpStatus.NOT_FOUND); }
         if(this.service.delete(id)) {
             return new ResponseEntity(HttpStatus.valueOf(200));
         } else {
